20220104 Second tutorial finished.  重點要留意︰1. Model.findByIdAndUpdate() 不等於 Query.prototype.findOneAndUpdate()  2. 唔好死係res.json，可以console + Network 睇下 listen 到啲backend object.

+++ server/controllers/post.js
@@ -34,9 +34,12 @@ export const updatePosts = async (req, res) => {
   const post = req.body;
 
   try {
-    if (!mongoose.Types.ObjectId.isValid(_id)) return res.status(404).send("No post with that id");
+    if (!mongoose.Types.ObjectId.isValid(_id))
+      return res.status(404).send("No post with that id");
 
-    let updatedPost = await PostMessage.findOneAndUpdate(_id, post, { new: true });
+    let updatedPost = await PostMessage.findOneAndUpdate(_id, post, {
+      new: true,
+    });
     res.status(200).json(updatedPost);
   } catch (err) {
     console.log(err);
@@ -47,8 +50,35 @@ export const deletePost = async (req, res) => {
   const { id: _id } = req.params;
 
   try {
+    if (!mongoose.Types.ObjectId.isValid(_id))
+      return res.status(404).send("No post with that id");
+
     const deletedPost = await PostMessage.deleteOne({ id: _id });
     res.json(`${deletedPost} has been deleted`);
+  } catch (err) {
+    console.log(err);
+  }
+};
+
+export const likePost = async (req, res) => {
+  const { id: _id } = req.params;
+  console.log(_id);
+
+  try {
+    if (!mongoose.Types.ObjectId.isValid(_id))
+      return res.status(404).send("No post with that id");
+
+    const post = await PostMessage.findById(_id);
+    // console.log(post._id);
+    // console.log(post);
+
+    const updatedPost = await PostMessage.findOneAndUpdate(
+      { _id: _id },
+      { likeCount: post.likeCount + 1 },
+      { new: true }
+    );
+    console.log(updatedPost._id);
+    res.json(updatedPost);
   } catch (err) {
     console.log(err);
   }
